requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.7.4
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-telegram-bot==21.4
pytz==2024.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
